requirements.txt
Django==4.2.11
djangorestframework==3.15.0
pytest==8.1.1
pytest-django==4.8.0
six==1.16.0
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
requests==2.28.2
urllib3==1.26.14
SecretStorage==3.3.3
pycryptodome==3.19.0